refactor(web): move use-autosave-queue to shared composables
The queue is consumed by both the bots pages and the providers page;
a provider importing a bot page's private components directory crosses
the documented feature layout (apps/web/AGENTS.md: shared composables
live in src/composables/). Pure move plus import rewrites, no behavior
change.
