+++ apps/web/src/pages/bots/components/bot-heartbeat.vue
@@ -313,7 +313,7 @@ import type { SettingsSettings, SettingsUpsertRequest, HeartbeatLog } from '@mem
 import { useQuery, useQueryCache } from '@pinia/colada'
 import { resolveApiErrorMessage } from '@/utils/api-error'
 import { formatDateTime } from '@/utils/date-time'
-import { useAutosaveQueue, type AutosaveJob } from './use-autosave-queue'
+import { useAutosaveQueue, type AutosaveJob } from '@/composables/use-autosave-queue'
 import type { Ref } from 'vue'
 
 const props = defineProps<{
